fix/kludge: Add support to read in port from .env files
This step unexpectedly faced a number of obstacles (I would have assumed
that setting the port would be a simple task). The main issues were that:
- the env file was not being loaded
- when running the script from package.json an old version of node (16.13.2) was being used (system version: 21.1.0)

In the end the solution was to write scripts to load the various run configs
explicitly. This is not ideal (as I was not able to identify the setup problem
yet), but it works (and arguably is the cleaner solution anyway).

ToDo:
- [ ] Investigate why the env file is not being loaded
- [ ] Investigate why the system node version is not being used
- [ ] Ensure that the change in .env has not resulted in any issues with sibling components (NestJS, Database, etc)