(i!) feat: Add login route with LocalAuthGuard for authentication
Milestone:
---------
You should now be able to log in (Postman)
  url: http://localhost:3000/auth/login
  method: POST
  body: {
    "email": [check database],
    "password": [check database]
  }

Mechanims:
---------
- Guards effectively represent the mechanism for integrating strategies into the controller
  -- The AuthGuard is generated by taking in the Passport module as an argument;
  -- A route that is decorated with an auth module triggers the corresponding strategy.

After defining the login route (controller) we "decorate it with at useGuards from Nestjs/Common and
pass in our local auth guard. This will ensure any time this route is called, we're going to run it
through the local auth guard.
And that's going to trigger the [...validate() method of the...] local strategy to run.