setup: add cookie-parser (to handle cookies in responses)
Context:
Now that we have a way to attach JWT to the response cookie (i.e. will be returned to server
from client on subsequent requests), we turn our attention to implementing a JWT strategy (i.e.
authenticate requests based on the JWT token).

Scope:
Install the cookie-parser library: middleware that:
- parses cookies of incoming requests (if authenticated will contain valid JWT token)
- populates the cookies object on the incoming request with any cookies
=> Makes it easy to access cookies in the request object