(!!) feat: (Major) Adding Signup functionality (full stack)
At this point you are able to start up both the NestJS server and the
NextJS Web App, browse to the signup page (http://localhost:3003/auth/signup
if started from npm run dev; http://localhost:3000/auth/signup if started from
the debugger), fill in the form and submit your sign up request. The NestJS
server will either create a new record in the database if the information is
valid or return an error response.

If the operation is successful the Web App will display a success message
(Toast) and redirect you to the login page. If the operation fails the Web App
will display an error message (Toast) and the app will remain on the signup
page.

Failed requests that have been tested include:
- Duplicate email address
- Passwords dont match

As already alluded to, we not only can start the application in dev mode
but also in the debugger.
-> To start the application in the debugger:
  1. Open 2x VSCode windows
    - Window 1: in core-ui
    - Window 2: in core-be
  2. click on the debugger icon in each window and start the debugger
  3. Select and start the run configuration in each window
    - Debug NestJS
    - (default) Official Next.js: debug full stack

-> To start the application in dev mode:
    1. Open a VSCode window in the project directory (~/MyProjects/Tutorials/Udemy_NestJS-NextJS_Guay_V2/testbench_Nest_Next_Guay)
    2. Open 2x terminals
    3.a Terminal 1:
      - cd core-ui
      - npm run dev
    3.b Terminal 2:
      - cd core-be
      - npm run start:dev