feat: add logout button (landing page) & functionality
After logging in you can see the auth token (be-core-auth) in the browser's
cookie jar (previous commit). The logout button will remove the token from the
cookie jar and redirect the user to the login page.

This is a pre-requisite step for testing access restrictions etc: if we cannot
logout we would have to manually delete cookies to test the login flow again.