(!!)feat: implement test bench for implementing the passing of (auth) cookies from the NextJS app to the NestJS server
Currently we can log into our application by sending a POST request to the
NestJS server (/auth/login) with the user's credentials. The server will
respond with a Set-Cookie header that contains the JWT token.

Since the NextJS back end server is making the request, however, the cookie
is not automatically stored in the browser. This means that the user is not
logged in when they are redirected to the home page. This functionality was
previously implemented in the NextJS app: Upon logging in, the browser stores
an auth JWT (be-core-auth) in the browser's cookies.

Now, when a new request is made to the NextJS server, this cookie containing
the auth JWT is attached - but the cookie is STILL not being passed to the
NestJS. This means that the server still does not know whether the user is
logged in or not.

To implement this functionality our landing page was modified to include a
check if the identity of the user is known. A protected route (whoami, GET)
was previously created on the NestJS server to return the details of the
current user (if known, i.e. if logged in). This information will be extracted
from the JWT token stored in the cookie. In other words, only if the JWT
is forwarded to the NestJS server will the server be able to determine the
identity of the user.

Attaching the JWT will be done in the next commit. Here we simply attempt
to check the identity of the user (i.e. attempt to find and read the JWT)
in NestJS. Since we are not passing the cookie containing the JWT to NestJS
yet this will always fail: The landing page is rendered with a message prompting
the user to sign in.