(!!) feat: Inform NestJS server about the auth status of the user in NextJS app
As outlined in the previous commit, in order for the NestJS server to
protects access to restricted routes (AuthGuard), it needs to know the
auth status of the user in the NextJS app. This is stored in the user's
browser as a cookie containing the jwt token. However, this cookie is
not automatically forwarded to the NestJS server when the user makes a request.

In this commit we explicitly forward the jwt token to the NestJS server
allowing it to establish the auth status of the user (ie manage access to
its routes).

------------------------------------------------------------------------
The ability to forward the jwt token to the NestJS server is a critical
feature as it allows the NestJS server to establish the auth status of the
user and thus manage (restrict) access to its routes.
Of note: if the JWT does not exist or is invalid, the NestJS server will
return a 401 Unauthorized response.
------------------------------------------------------------------------

In the process we also implemented the whoami/ route which returns the
user account details of the logged in user. This enables user-specific
functionalities to also be implemented on the web app - e.g.
- Personalization (e.g. display user's name in the header, greeting messages etc)
- Conditional rendering of components (e.g. display logout button only if user is logged in)
- Conditional (app) routing (e.g. redirect user to login page if not logged in)
- etc.

Test workflow:
-> Start up the application - you should be redirected to the landing page with
   a message to log in.
-> If your browse to the login page and provide valid credentials you will
    be redirected to the landing page - but this time a logout button is displayed
-> If you click the logout button the landing page will again display the
    message to log in.
redirections are not fully implemented to allow you to test the login/logout
functionality at will.