(!!) feat: Implementing middleware for access protecting NextJS routes
NextJS web apps have the app router (alternatively "pages router") to
implement routing for the app. We want to protect some routes from
unauthorized access.

There are different strategies for implementing access protection - one
is the implementation of middleware functions. Middleware functions are
functions that have access to the request object (req), the response
object (res), and the next middleware function in the application’s
request-response cycle.

NextJS middleware must be in a root level file called "middleware.ts".

We explicitly state public routes (all other routes are access protected).
We define a config object that contains exceptions (i.e. resources that
should be public based on criteria other than URL's defined as being public).
This was defined for static resources.

The way the Middleware works is that it checks if the route is public, or
if it has been excluded on account of being added to the config object.
If neither of these exclusions apply, the middleware checks if the user
is authenticated (i.e has an auth cookie - NB: it does not check if the cookie
is valid (!)

If the resource is not public, and the user is not authenticated, the user
is redirected to the login screen to provide credentials. If the credentials
are valid, the user is automatically redirected back to the requested resource.