debug: Update launch configuration names for consistency and improve script path resolution for Next.js
Updating the run scripts was necessary as the application could not be run from the terminal using: npm start dev. This works now.
Updating the launch configuration names was an attempt to simplify them - as the "Problems" tab in VSCode was showing a warning that the names were not unique.
  This change had no visible effect (but was kept in in as the simplified names are preferred). The warning message has yet to be resolved!!